requirements.txt
asgiref==3.2.7
astroid==2.4.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.9
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.4
openapi-codec==1.3.2
pandas==1.0.3
pylint==2.5.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
